- Support for loading plugins from a theme
A major update which allows theme authors to bring plugins with
their themes.

Additionally to this, the logic around protected options was
enhanced.  Now the protected variables are exposed to the
theme configuration files, so it is possible to do something
like the following in `themeconf.py`:

PLUGIN_PATH.append(THEME + '/plugins')